Refactor of code - cleanup and more
