+++ configuration.properties
@@ -1,4 +1,8 @@
 # Configuration for Nebula
+throttling_delay=250
+max_accounts=5
+connect_message=
+server_history=astropvp.net, cosmicpvp.com, hypixel.net
 
 open_in_new_terminal=false
 account_configuration_name=account_configurations.csv
